semantic-search/requirements.txt
endee>=0.1.16
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=1.0.0
